Midterm/e_commerce/requirements.txt
asgiref==3.8.1
Django==5.1.2
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-phonenumber-field==8.0.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
gunicorn==23.0.0
inflection==0.5.1
packaging==24.1
phonenumberslite==8.13.47
psycopg2==2.9.10
PyJWT==2.9.0
pytz==2024.2
PyYAML==6.0.2
redis==5.1.1
sqlparse==0.5.1
uritemplate==4.1.1
